requirements.txt
python-dotenv
requests
requests-mock
